requirements.txt
click==6.7
colorama==0.3.9
python-dotenv==0.13.0
streamlit
